Auto-Initialise XPlotly.Plotly.fsx and FSCharting.fsx (#84)
* Auto-Initialise FSCharting.fsx

This change automatically initialises Display printers and uses
[<AutoOpen>] to expose the extension methods. As a result, the
notebook user only needs to #load the helper script, and doesn't
need to know a module name to open or an initialisation function
to run.

* Auto-initialise XPlot.Plotly.fsx

This change uses `addDisplayPrinter` to make XPlot charts display
in the same way as other elements. It also changes the way the
plot.ly JavaScript is loaded such that it is loaded in the notebook
browser when the script is loaded, without needing an initialisation
call.

* Allow XPlot.Plotly charts to render as images

Plotly has many lovely features, but large data sets rendered as
JavaScript can be difficult to deal with. This adds Chart.Image,
which renders a chart as a static PNG.

* Consistently load plotly JS across browsers

* Chart.Image generates SVG on Microsoft browsers, as PNG generation fails