Install option (#87)
Previously, ifsharp.exe would always try to run jupyter. This is not
the desired behaviour in some instances (for example, in a Docker
container, where jupyter must be started with specific command line
options).

This change adds a `--install` option that installs the IfSharp
kernel without trying to start jupyter.