Paket based dependencies
Rather than depend on a fixed set of packages, use Paket to fetch all
dependencies dynamically.

Helper scripts are divided into two parts to allow Paket to run _before_
the `#r` directives in the second script are evaluated. This is necessary
because preprocessor directives are evaluated before all code, and this
appears to be transitive through `#load` directives - that is, a wrapper
scripts that loads both the Paket script and the helper script also
evaluates the `#r` directives in the second script before code in the
first script is run.