Syntax and intellisense bug fix simplification
There were nasty race conditions in the way we were loading CodeMirror
and setting our language see #135 I've simplified it and enforced an
ordering. This means we don't have to force CodeMirror so much but we
still get predictable intellisense. It means syntax highlighting doesn't
apply so soon but it reliably works!