Fake with current .NET Core runner (#238)
Moved to .NET Core based Fake, replacing the compatiblity mode we had
Experimental .NET Core builds on macOS, Ubuntu, and Windows on Azure DevOps
The Mono/.NET Framework worked for Windows but not Ubuntu on DevOps, so keeping Travis as well
The Travis build was also running a Docker deploy, which on a branch doesn't totally make sense. I've disabled that for now, volunteers would be great.